Add files via upload
Access Control Lists (ACLs) are a fundamental component of network security used to control the flow of traffic into and out of a network. An ACL is a set of rules that are applied to network devices, such as routers and firewalls, to permit or deny traffic based on various criteria such as IP addresses, protocol types, and port numbers.